requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==24.2.0
beautifulsoup4==4.12.3
blendmodes==2022
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
coloredlogs==15.0.1
colorlog==6.8.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.11
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1680755705686/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deprecation==2.1.0
depth_anything @ https://github.com/huchenlei/Depth-Anything/releases/download/v1.0.0/depth_anything-2024.1.22.0-py2.py3-none-any.whl#sha256=26c1d38b8c3c306b4a2197d725a4b989ff65f7ebcf4fb5a96a1b6db7fbd56780
depth_anything_v2 @ https://github.com/MackinationsAi/UDAV2-ControlNet/releases/download/v1.0.0/depth_anything_v2-2024.7.1.0-py2.py3-none-any.whl#sha256=6848128867d1f7c7519d88df0f88bfab89100dc5225259c4d7cb90325c308c9f
diffusers==0.30.3
diskcache==5.6.3
easydict==1.13
ecdsa==0.19.0
einops==0.4.1
embreex==2.17.7.post5
eval_type_backport==0.2.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
facexlib==0.3.0
fastapi==0.115.3
fastprogress==1.0.3
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.32
gradio==5.3.0
gradio_client==1.4.2
gradio_imageslider==0.0.20
gradio_rangeslider==0.0.6
h11==0.12.0
handrefinerportable @ https://github.com/huchenlei/HandRefinerPortable/releases/download/v1.0.1/handrefinerportable-2024.2.12.0-py2.py3-none-any.whl#sha256=1e6c702905919f4c49bcb2db7b20d334e8458a7555cd57630600584ec38ca6a9
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
ifnude==0.0.3
imageio==2.36.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
inflection==0.5.1
insightface==0.7.3
iopath==0.1.9
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1727944696411/work
jax==0.4.34
jaxlib==0.4.34
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonmerge==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.7
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
lightning-utilities==0.11.7
llvmlite==0.43.0
loadimg==0.1.2
lxml==5.3.0
manifold3d==2.5.1
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mediapipe==0.10.15
ml_dtypes==0.5.0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.4.1
numba==0.60.0
numpy==1.26.2
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.19.2
open-clip-torch==2.20.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.12.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pooch==1.8.2
portalocker==2.10.1
prettytable==3.11.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.0
protobuf==4.25.5
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1728965182690/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyasn1==0.6.1
pycollada==0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyMatting==1.1.12
pyparsing==3.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-jose==3.3.0
python-multipart==0.0.12
pytorch-lightning==1.9.4
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1698062577979/work
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
reportlab==4.2.5
requests==2.32.3
resize-right==0.0.2
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
Rtree==1.3.0
ruff==0.6.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spandrel==0.3.4
spandrel_extra_arches==0.1.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.41.0
svg.path==6.3
svglib==1.5.1
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.20.1
tomesd==0.1.3
tomlkit==0.12.0
torch==2.2.2
torchaudio==2.2.2
torchdiffeq==0.2.3
torchmetrics==1.4.3
torchsde==0.2.6
torchvision==0.17.2
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1724956089949/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
trampoline==0.1.2
transformers==4.45.2
trimesh==4.5.0
typer==0.12.5
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
vhacdx==0.0.8.post1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
websockets==12.0
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
yarl==1.15.2
yt-dlp==2024.10.7
zipfile36==0.1.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
